requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.42.3
torch==2.3.1
Pillow==10.4.0
requests==2.32.3
matplotlib==3.7.5
chainlit==2.6.0
pydantic==2.10.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
einops==0.8.0
timm==1.0.7
sniffio
anyio
boto3
sqlalchemy
greenlet
asyncpg
structlog
logfire
logfire[fastapi]
opentelemetry-instrumentation-celery